+++ pixelperfectengine/src/PixelPerfectEngine/graphics/layers/trnstilelayer.d
@@ -475,10 +475,10 @@ public class TransformableTileLayer(BMPType = Bitmap16Bit, int TileX = 8, int Ti
 	public int getMY() @nogc @safe pure nothrow const {
 		return mY;
 	}
-	public int getTX() @nogc @safe pure nothrow const {
+	public size_t getTX() @nogc @safe pure nothrow const {
 		return totalX;
 	}
-	public int getTY() @nogc @safe pure nothrow const {
+	public size_t getTY() @nogc @safe pure nothrow const {
 		return totalY;
 	}
 	/// Sets the warp mode.
